Using SDP Munging to identify connections.
Before I was using a signed addresses message to identify what peer was on the other side of a rtcpeerconnection, but now we store a public key and signature replacing the SDP session name. (The session name is arbitrary and browsers seem to just fill it with a single "-" character.  Honestly, not sure why it's still a required line in sdp.